+++ telekit/example/example_handlers/__init__.py
@@ -2,3 +2,4 @@ from . import start
 from . import entry
 from . import help
 from . import on_text
+from . import faq