+++ CHANGELOG.md
@@ -1,14 +1,15 @@
 # 0.2.0
 ## New Features
+- `telekit.GuideMixin` and `telekit.GuideKit(...).register()`: allows creating interactive FAQ pages using a custom DSL, with automatic scene handling, user input processing, and message formatting.
 - `@chain.entry_location()` decorator: allows handling messages with user coordinates.
 - `@chain.on_timeout()` decorator: registers a callback to be executed after a timeout.  
 - `chain.set_timeout()` an alternative way to set a timeout programmatically.
+- `server.long_polling()` an alternative to `server.polling()` that uses long polling with a configurable timeout
 
 ## Planned:
+- Result caching in `GuideKit`
 - Ability to disable logging `server.enable_logging(True)`
 - Localization of the method effect `self.user.enable_logging()` (Currently working globally)
-- New listeners
-- New ways to "polling"?
 
 ---